requirements.txt
flask
kafka-python
psycopg2
json
